Modularize language bundles for i18n
