+++ assets/i18n/languages.json
@@ -1,0 +1,5 @@
+[
+  { "value": "en", "label": "English", "module": "./en.js" },
+  { "value": "zh", "label": "简体中文", "module": "./zh.js" },
+  { "value": "ja", "label": "日本語", "module": "./ja.js" }
+]
